Update to ares v112r04 release.
  - N64: merged the VR4300 core into the CPU class
  - mia: added N64 deinterleaving support

+++ ares/n64/cartridge/cartridge.hpp
@@ -1,6 +1,6 @@
 struct Cartridge {
   Node::Peripheral node;
-  Memory::Readable<uint8> rom;
+  Memory::Readable<uint32> rom;
 
   auto manifest() const -> string { return information.manifest; }
   auto name() const -> string { return information.name; }
